refactor(models): consolidate duplicate room, item and team models
Three tables were mapped by two model classes each. Keep the richer
class for every table: App\Models\Room absorbs the guild and owner
relations from the deleted Game\Room (which lacked timestamps handling),
ItemBase replaces the docblock-only ItemDefinition clone, and the two
WebsiteTeam classes merge into Community\Teams\WebsiteTeam, which now
implements HasBadge. All imports, the observer registrations and the
dashboard widget point at the surviving classes, and Room::replicateForUser
now runs inside a transaction. The EventServiceProvider also registers
the new WebsiteAd and WebsiteOpenPosition observers.
