+++ app/Filament/Resources/Atom/Teams/TeamResource.php
@@ -7,7 +7,7 @@ use App\Filament\Resources\Atom\Teams\Pages\EditTeam;
 use App\Filament\Resources\Atom\Teams\Pages\ListTeams;
 use App\Filament\Tables\Columns\HabboBadgeColumn;
 use App\Filament\Traits\TranslatableResource;
-use App\Models\Community\Staff\WebsiteTeam;
+use App\Models\Community\Teams\WebsiteTeam;
 use Filament\Actions\DeleteBulkAction;
 use Filament\Actions\EditAction;
 use Filament\Forms\Components\TextInput;
